test(api): split embedding administration schema coverage
